refactor: moved docs
